todoTask v1

+++ src/main/java/com/luiamerico/todoapplication/service/TaskService.java
@@ -2,7 +2,9 @@ package com.luiamerico.todoapplication.service;
 
 import com.luiamerico.todoapplication.model.Task;
 import com.luiamerico.todoapplication.repository.TaskRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -17,12 +19,29 @@ public class TaskService {
     public List<Task> getAllTasks() {
         return taskRepository.findAll();
     }
-    
+
+    public Task getTaskById(Long id) {
+        return taskRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found"));
+    }
+
+    public Task updateTask(Long id, Task task) {;
+        var existingTask = taskRepository.findById(id);
+        if (existingTask.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+        task.setId(id);
+        return taskRepository.save(task);
+    }
+
     public Task createTask(Task task) {
         return taskRepository.save(task);
     }
 
     public void deleteTask(Long id) {
+        if (!taskRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Task not found");
+        }
         taskRepository.deleteById(id);
     }
